Added Test and MultiHeadOutput
