+++ tests/test_adapter.py
@@ -7,7 +7,7 @@ from transformers import BartConfig, BertConfig, DistilBertConfig, GPT2Config, M
 from transformers.testing_utils import require_torch, torch_device
 
 from .test_adapter_common import AdapterModelTestMixin
-from .test_adapter_composition import ParallelAdapterInferenceTestMixin, ParallelTrainingMixin
+from .test_adapter_composition import ParallelAdapterInferenceTestMixin
 from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
@@ -55,7 +55,6 @@ class BertAdapterTest(
     PredictionHeadModelTestMixin,
     AdapterTrainingTestMixin,
     ParallelAdapterInferenceTestMixin,
-    ParallelTrainingMixin,
     BertAdapterTestBase,
     unittest.TestCase,
 ):
